+++ R/az_subscription.R
@@ -7,12 +7,12 @@
 #' @docType class
 #' @section Methods:
 #' - `new(token, id, ...)`: Initialize a subscription object.
-#' - `list_resource_groups()`: Return a list of resource group objects for this subscription.
+#' - `list_resource_groups(filter, top)`: Return a list of resource group objects for this subscription. `filter` and `top` are optional arguments to filter the results; see the [Azure documentation](https://docs.microsoft.com/en-us/rest/api/resources/resourcegroups/list) for more details. If `top` is specified, the returned list will have a maximum of this many items.
 #' - `get_resource_group(name)`: Return an object representing an existing resource group.
 #' - `create_resource_group(name, location)`: Create a new resource group in the specified region/location, and return an object representing it. By default, AzureRMR will set the `createdBy` tag on a newly-created resource group to the value `AzureR/AzureRMR`.
 #' - `delete_resource_group(name, confirm=TRUE)`: Delete a resource group, after asking for confirmation.
 #' - `resource_group_exists(name)`: Check if a resource group exists.
-#' - `list_resources()`: List all resources deployed under this subscription.
+#' - `list_resources(filter, expand, top)`: List all resources deployed under this subscription. `filter`, `expand` and `top` are optional arguments to filter the results; see the [Azure documentation](https://docs.microsoft.com/en-us/rest/api/resources/resources/list) for more details. If `top` is specified, the returned list will have a maximum of this many items.
 #' - `list_locations()`: List locations available.
 #' - `get_provider_api_version(provider, type, which=1, stable_only=TRUE)`: Get the current API version for the given resource provider and type. If no resource type is supplied, returns a vector of API versions, one for each resource type for the given provider. If neither provider nor type is supplied, returns the API versions for all resources and providers. Set `stable_only=FALSE` to allow preview APIs to be returned. Set `which` to a number > 1 to return an API other than the most recent.
 #' - `do_operation(...)`: Carry out an operation. See 'Operations' for more details.
@@ -157,11 +157,16 @@ public=list(
         az_resource_group$new(self$token, self$id, name)
     },
 
-    list_resource_groups=function()
+    list_resource_groups=function(filter=NULL, top=NULL)
     {
-        cont <- private$sub_op("resourcegroups")
-        lst <- lapply(get_paged_list(cont, self$token), function(parms)
-            az_resource_group$new(self$token, self$id, parms=parms))
+        opts <- list(`$filter`=filter, `$top`=top)
+        cont <- private$sub_op("resourcegroups", options=opts)
+        lst <- lapply(
+            if(is.null(top))
+                get_paged_list(cont, self$token)
+            else cont$value,
+            function(parms) az_resource_group$new(self$token, self$id, parms=parms)
+        )
 
         named_list(lst)
     },
@@ -185,11 +190,16 @@ public=list(
         httr::status_code(res) < 300
     },
 
-    list_resources=function()
+    list_resources=function(filter=NULL, expand=NULL, top=NULL)
     {
-        cont <- private$sub_op("resources")
-        lst <- lapply(get_paged_list(cont, self$token), function(parms)
-            az_resource$new(self$token, self$id, deployed_properties=parms))
+        opts <- list(`$filter`=filter, `$expand`=expand, `$top`=top)
+        cont <- private$sub_op("resources", options=opts)
+        lst <- lapply(
+            if(is.null(top))
+                get_paged_list(cont, self$token)
+            else cont$value,
+            function(parms) az_resource$new(self$token, self$id, deployed_properties=parms)
+        )
 
         names(lst) <- sapply(lst, function(x) sub("^.+providers/(.+$)", "\\1", x$id))
         lst