+++ tests/testthat/test05b_template.R
@@ -75,6 +75,16 @@ test_that("Template methods work",
 
     # tagging
     expect_identical(tpl4$get_tags(), list(createdBy="AzureR/AzureRMR"))
+
+    # listing
+    tpllst0 <- rg$list_templates()
+    expect_true(is.list(tpllst0) && all(sapply(tpllst0, is_template)))
+
+    tpllst <- rg$list_templates(top=1)
+    expect_true(is.list(tpllst) && length(tpllst) == 1)
+
+    tpllst <- rg$list_templates(filter="provisioningState eq 'Succeeded'")
+    expect_true(is.list(tpllst) && length(tpllst) > 0)
 })
 
 rg$delete(confirm=FALSE)