+++ NEWS.md
@@ -1,5 +1,6 @@
 # AzureRMR 2.3.3.9000
 
+- Add filtering arguments (`filter`, `expand` and `top`) for the `list_resource_groups`, `list_templates` and `list_resources` methods, to trim the results. See the Azure docs for more details.
 - Add `createdBy:AzureR/AzureRMR` tag to Azure objects (resource groups, resources and templates) created by this package.
 - Add a `get_tags()` method for templates.